requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asyncmy==0.2.9
certifi==2024.8.30
click==8.1.7
databases==0.9.0
dlib==19.24.6
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.115.4
fastapi-cli==0.0.5
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.1.2
opencv-python==4.10.0.84
pillow==11.0.0
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyMySQL==1.1.1
pysnmp==7.1.13
python-dotenv==1.0.1
python-multipart==0.0.16
PyYAML==6.0.2
rich==13.9.3
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
typer==0.12.5
typing_extensions==4.12.2
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
